ProtoViT/requirements.txt
absl-py==2.1.0
Augmentor==0.2.10
blessed==1.20.0
cloudpickle==3.1.0
contourpy==1.1.1
cmake>=3.18
cleverhans
decorator==5.1.1
dm-tree==0.1.8
easydict==1.13
fsspec==2023.12.2
gast==0.6.0
gpustat==1.1.1
huggingface-hub==0.20.3
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
lit>=16.0.0
matplotlib==3.7.4
mnist==0.2.2
networkx==3.1
ninja==1.11.1.1
nose==1.3.7
nvidia-ml-py==12.535.133
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
psutil==5.9.8
pycodestyle==2.12.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
safetensors==0.4.2
scikit-learn==1.3.2
six==1.16.0
tensorflow-probability==0.21.0
threadpoolctl==3.2.0
timm
tokenizers
tqdm
transformers
tzdata==2023.4
wcwidth==0.2.13
zipp==3.17.0